Working on object scale fixes
This is a WIP. When I tell a bubble to scale to 10, I want it to be 10
units wide. This fixes that. It also fixes the fresnel shader, which i'm
guessing just worked by luck before.

Still to fix: tweaking fresnel values ( doesn't show ripples at small
size ), size of mines, and distortion of fireballs