Introducing spawner object for bubble spawn position
Adding a spawner invisible 3d object which is essentially a cube in
which bubbles spawn per level, defined per level. The default cube is 0
high, 0 deep, and camera.width wide, and placed at the top of the
screen, so bubbles are essentially spawning along a line.

I think the api for the spawner and the mode access to the spawner could
be greatly improved. I am certainly ashamed of how I handled camera
zooming in this commit.